+++ lib/src/hierarchy.rs
@@ -30,7 +30,7 @@ impl fmt::Display for Right {
     }
 }
 
-/// Looks for children relations, adds to the resource. Performs a TPF query, might be expensive.
+/// Looks for children relations, adds to the resource. Performs a Query, might be expensive.
 pub fn add_children(store: &impl Storelike, resource: &mut Resource) -> AtomicResult<Resource> {
     let results = store.query(&Query::new_prop_val(urls::PARENT, resource.get_subject()))?;
     let mut children = results.subjects;