+++ CHANGELOG.md
@@ -5,6 +5,10 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 Changes to JS assets are not included here, but in [`atomic-data-browser`'s CHANGELOG](https://github.com/atomicdata-dev/atomic-data-browser/blob/main/CHANGELOG.md).
 See [STATUS.md](server/STATUS.md) to learn more about which features will remain stable.
 
+## UNRELEASED
+
+- Remove `tpf` queries from `atomic-cli` #610
+
 ## [v0.34.2] - 2023-03-04
 
 - **Requires `--rebuild-index`**